Split app startup bootstrap out of app-init
